% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AvailableData}
\alias{AvailableData}
\title{List available datasets}
\usage{
AvailableData()
}
\value{
A data frame with columns \code{name} and \code{description}
}
\description{
Returns a table of dataset names and descriptions
}
