% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ConvertGenes}
\alias{ConvertGenes}
\title{Convert genes}
\usage{
ConvertGenes(fingerprints, from = "ENSEMBL", to = "SYMBOL")
}
\arguments{
\item{fingerprints}{Fingerprints obtained from LearnDictionary()}

\item{from}{Key type to convert from (default 'ENSEMBL')}

\item{to}{Key type to convert to (default 'SYMBOL')}
}
\value{
Fingerprints with re-named genes; genes that couldn't be converted are dropped
}
\description{
Convert genes in a fingerprints object
}
\examples{
\dontrun{
fingerprints <- ConvertGenes(
fingerprints
)
}

}
