% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{InstallPrecomputedFingerprints}
\alias{InstallPrecomputedFingerprints}
\title{Install precomputed fingerprints}
\usage{
InstallPrecomputedFingerprints(name)
}
\arguments{
\item{name}{Name of dataset; see available datasets with AvailableData()}
}
\value{
No return value; installs requested fingerprints
}
\description{
Installs requested precomputed fingerprints
}
