% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{LoadPrecomputedFingerprints}
\alias{LoadPrecomputedFingerprints}
\title{Load pre-computed fingerprints}
\usage{
LoadPrecomputedFingerprints(name, gene_key = "ENSEMBL")
}
\arguments{
\item{name}{Name of pre-computed fingerprints}

\item{gene_key}{Formatting of genes (default ENSEMBL); ignored for mouse datasets}
}
\value{
A set of pre-computed fingerprints
}
\description{
Load in pre-computed fingerprints.
}
\examples{
\dontrun{
gwps_fingerprints <- LoadPrecomputedFingerprints(
name = 'gwps'
)

gwps_fingerprints <- LoadPrecomputedFingerprints(
name = 'gwps',
gene_key = 'SYMBOL'
)
}

}
